Move CellContent component, avoid duplicate HeaderValuePair type
