+++ frontend/src/components/pages/ecommerce/SalesTermsDialog.tsx
