+++ frontend/src/interfaces/utils.ts
@@ -1,0 +1,5 @@
+export type HeaderValuePair<T> = {
+    header: string;
+    field: keyof T;
+  };
+
